+++ frontend/src/mod/algorithm/parserTools/test/parser.test.ts
@@ -1,9 +1,9 @@
 import { expect, test } from 'vitest';
 import { Polygon } from '@/mod/data/polygon/polygon';
 import { Room } from '@/mod/data/room/room';
-import * as polygonTools from './polygonTools';
-import * as roomTools from './roomTools';
-import * as graphTools from './graphTools';
+import * as polygonTools from '../polygonTools';
+import * as roomTools from '../roomTools';
+import * as graphTools from '../graphTools';
 
 import type { Line } from '@/mod/data/com/line';
